requirements.txt
Flask==3.0.0
bcrypt==4.1.2
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
itsdangerous==2.1.2
psycopg2-binary==2.9.9